+++ R/get_solution.R
@@ -1,15 +1,11 @@
 #' Get an Astronomical Solution
 #'
-#' @param orbital_solution Character vector with the name of the orbital
-#'   solution to use or a `data.frame` with a custom orbital solution.
-#'   One of `"ZB18a"` (default) from Zeebe and Lourens (2019),
-#'   or `"La11"` (not yet implemented!).
-# quiet and force are documented in get_ZB18a
-#' @inheritParams get_ZB18a
-#' @seealso [get_ZB18a()]
-#' @inherit get_ZB18a references
-#' @returns `get_solution()` returns a [tibble][tibble::tibble-package] with
-#'   the orbital solution input and some preprocessed new columns.
+# astronomical_solution, quiet, and force are documented in get_ZB
+#' @inheritParams get_ZB
+#' @seealso [get_ZB()]
+#' @inherit ZB18a references
+#' @returns A [tibble][tibble::tibble-package] with the astronomical solution
+#'   (and some preprocessed new columns).
 #' @examples
 #' \donttest{
 #' get_solution()
@@ -20,15 +16,19 @@ get_solution <- function(astronomical_solution = "PT-ZB18a", quiet = FALSE, forc
     return(prepare_solution(astronomical_solution, quiet = quiet))
   }
 
-  solutions <- c("ZB18a", "La11")
-  if (!orbital_solution %in% solutions) {
-    cli::cli_abort(c("{.var orbital_solution} must be one of: {.or {.q {solutions}}}",
-                     "x" = "You've supplied {.q {orbital_solution}}"))
-  }
+  # effectively this is a wrapper that checks only for valid solutions...
+  solutions <- c("PT-ZB18a", # the default
+                 "PT-La11", # just to thow an error
+                 "ZB17a", "ZB17b", "ZB17c", "ZB17d",
+                 "ZB17e", "ZB17f", "ZB17h", "ZB17i",
+                 "ZB17j", "ZB17k", "ZB17p",
+                 "ZB18a-100",
+                 "ZB18a-300",
+                 "ZB20a", "ZB20b", "ZB20c", "ZB20d")
 
-  if (orbital_solution == "ZB18a") {
-    # read in the (new?) cached result
-    dat <- get_ZB18a(quiet = quiet, force = force)
+  if (!astronomical_solution %in% solutions) {
+    cli::cli_abort(c("{.var astronomical_solution} must be one of: {.or {.q {solutions}}}",
+                     "x" = "You've supplied {.q {astronomical_solution}}"))
   }
 
   if (astronomical_solution == "PT-La11") {
@@ -42,6 +42,11 @@ get_solution <- function(astronomical_solution = "PT-ZB18a", quiet = FALSE, forc
                      "i" = "Pull requests welcome."))
   }
 
+
+  if (astronomical_solution == "PT-ZB18a" ||
+        stringr::str_detect(astronomical_solution, "^ZB[0-9][0-9][a-z]")) {
+    dat <- get_ZB(astronomical_solution, quiet = quiet, force = force)
+  }
 
   return(dat)
 }
@@ -57,54 +62,77 @@ get_solution <- function(astronomical_solution = "PT-ZB18a", quiet = FALSE, forc
 #'   * If `FALSE` (the default) print info messages and timing.
 #' @param force Force re-downloading the results, even if the solution is saved
 #'   to the cache.
-#' @returns `get_ZB18a()` returns a [tibble][tibble::tibble-package] with the
-#'   orbital solution input and some preprocessed new columns.
-#' @rdname ZB18a
-#' @seealso [prepare_solution()] Processes orbital solution input to include
-#'   helper columns.
+#' @returns A [tibble][tibble::tibble-package] with the astronomical solution
+#'   input and, in the case of the PT-ZB18a, some preprocessed new columns.
+#' @seealso [prepare_solution()] Processes precession-tilt astronomical
+#'   solution input to include helper columns.
 #' @examples
 #' \donttest{
-#' get_ZB18a()
+#' get_ZBa()
 #' }
-#' @export
-get_ZB18a <- function(quiet = FALSE, force = FALSE) {
-  ZB18a_url <- "http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat"
-
+get_ZB <- function(astronomical_solution = "PT-ZB18a",
+                   quiet = FALSE,
+                   force = FALSE) {
+  base_url <- "http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/"
   cachedir <- tools::R_user_dir("snvecR", which = "cache")
 
-  raw_path <- paste0(cachedir, "/ems-plan3.dat")
-  csv_path <- paste0(cachedir, "/ZB18a.csv")
-  rds_path <- paste0(cachedir, "/ZB18a.rds")
+  if (astronomical_solution == "PT-ZB18a") {
+    url <- paste0(base_url, "PrecTilt/OS/ZB18a/ems-plan3.dat")
+    # TODO: rename to PrecTilt, same as on ZB's website
+    raw_col_names <- c("t", # time in days
+                       "aa", # semimajor axis
+                       "ee", # eccentricity
+                       "inc", # inclination
+                       "lph", # long periapse
+                       "lan", # long ascending node
+                       "arp", # argument of periapse
+                       "mna") # mean anomaly
+  } else {
+    raw_col_names <- c("age", # time in ka, negative
+                       "ecc", # eccentricity
+                       "inc") # inclination
+  }
+  # different URLs for different solutions
+  if (stringr::str_detect(astronomical_solution, "^ZB1[78][a-z](-100)?$")) {
+    # 17 and 18 are stored in the root directory
+    url <- glue::glue("{base_url}{stringr::str_replace(astronomical_solution, '-100', '')}.dat")
+  } else if (stringr::str_detect(astronomical_solution, "^ZB[0-9][0-9][a-z](-300)?$")){
+    # ZB20 is stored in 300Myr subdirectory
+    url <- glue::glue("{base_url}300Myr/{stringr::str_replace(astronomical_solution, '-300', '')}.dat")
+  }
 
-  raw_col_names <- c("t", # time in days
-                     "aa", # semimajor axis
-                     "ee", # eccentricity
-                     "inc", # inclination
-                     "lph", # long periapse
-                     "lan", # long ascending node
-                     "arp", # argument of periapse
-                     "mna") # mean anomaly
+  # where will we save our cached results?
+  raw_path <- paste0(cachedir, "/", astronomical_solution, ".dat")
+  csv_path <- paste0(cachedir, "/", astronomical_solution, ".csv")
+  rds_path <- paste0(cachedir, "/", astronomical_solution, ".rds")
 
   # read final processed file from cache if available
   if (!force && file.exists(rds_path)) {
-    ZB18a <- readr::read_rds(rds_path)
-    return(ZB18a)
+    rds <- readr::read_rds(rds_path)
+    return(rds)
   }
 
   # read raw intermediate steps from disk
   if (!force && (file.exists(csv_path) || file.exists(raw_path))) {
     if (file.exists(csv_path)) {
-      ZB18a_raw <- readr::read_csv(csv_path, show_col_types = FALSE)
+      raw <- readr::read_csv(csv_path, show_col_types = FALSE)
     }
 
     if (file.exists(raw_path)) {
-      ZB18a_raw <- readr::read_table(raw_path,
-                                     col_names = raw_col_names,
-                                     skip = 3, comment = "#",
-                                     show_col_types = FALSE)
+      if (astronomical_solution == "PT-ZB18a") {
+        raw <- readr::read_table(raw_path,
+                                 col_names = raw_col_names,
+                                 skip = 3, comment = "#",
+                                 show_col_types = FALSE)
+      } else {
+        raw <- readr::read_table(raw_path,
+                                 col_names = raw_col_names,
+                                 comment = "%",
+                                 show_col_types = FALSE)
+      }
     }
   } else {# files don't exist or force
-    if (!quiet) cli::cli_alert_info("The orbital solution ZB18a has not been downloaded.")
+    if (!quiet) cli::cli_alert_info("The astronomical solution {astronomical_solution} has not been downloaded.")
 
     # default to downloading/caching if not interactive (i.e. GitHub actions)
     if (force || !interactive()) {
@@ -122,23 +150,29 @@ get_ZB18a <- function(quiet = FALSE, force = FALSE) {
       }
     }
 
-    # the user would NOT like to download and process the orbital solution
+    # the user would NOT like to download and process the astronomical solution
     if (!download) {
-      cli::cli_abort("Cannot `get_ZB18a()` without downloading the file.")
+      cli::cli_abort("Cannot `get_ZB()` without downloading the file.")
     }
 
     # read the file from the website
-    ZB18a_raw <- readr::read_table(ZB18a_url,
-                                   col_names = raw_col_names,
-                                   skip = 3, comment = "#",
-                                   show_col_types = FALSE)
-    if (!quiet) cli::cli_alert_info("Read {.file {basename(raw_path)}} from website {.url {ZB18a_url}}.")
-
-    # calculate helper columns
-    ZB18a <- ZB18a_raw |> prepare_solution(quiet = quiet)
+    if (!quiet) cli::cli_alert_info("Reading {.file {basename(raw_path)}} from website {.url {url}}.")
+    if (astronomical_solution == "PT-ZB18a") {
+      raw <- readr::read_table(url,
+                               col_names = raw_col_names,
+                               skip = 3, comment = "#",
+                               show_col_types = FALSE)
+      # calculate helper columns
+      raw <- raw |> prepare_solution(quiet = quiet)
+    } else {
+      raw <- readr::read_table(url,
+                               comment = "%",
+                               col_names = raw_col_names,
+                               show_col_types = FALSE)
+    }
 
     if (!save_cache) {
-      return(ZB18a)
+      return(raw)
     } else {
       if (!dir.exists(cachedir)) {
         dir.create(cachedir, recursive = TRUE, showWarnings = TRUE)
@@ -147,20 +181,20 @@ get_ZB18a <- function(quiet = FALSE, force = FALSE) {
 
       # also copy the raw file to disk
       # even though we've read it in using read_table directly
-      curl::curl_download(ZB18a_url, destfile = raw_path)
+      curl::curl_download(url, destfile = raw_path)
       if (!quiet) cli::cli_alert_info("Saved {.file {basename(raw_path)}} to cache.")
 
       # write intermediate result to csv
-      readr::write_csv(ZB18a_raw, csv_path)
+      readr::write_csv(raw, csv_path)
       if (!quiet) cli::cli_alert_info("Saved cleaned-up {.file {basename(csv_path)}} to cache.")
 
       # write final result to rds cache
-      readr::write_rds(ZB18a, rds_path)
+      readr::write_rds(raw, rds_path)
       if (!quiet) {
         cli::cli_alert("Saved solution with helper columns {.file {basename(rds_path)}} to cache.",
                        "i" = "Future runs will read from the cache, unless you specify `force = TRUE`.")
       }
     }
-    return(ZB18a)
+    return(raw)
   }
 }