+++ data-raw/ZB18a.R
@@ -99,4 +99,4 @@ usethis::use_data(ZB18a,
 
 rm(dat)
 
-# all of this is now a part of the fuction get_ZB18a()!
+# all of this is now a part of the fuction get_ZB